// apps/web/src/shared/api/generated/model/listDetailsControllerDetailBookType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type ListDetailsControllerDetailBookType =
  (typeof ListDetailsControllerDetailBookType)[keyof typeof ListDetailsControllerDetailBookType];

export const ListDetailsControllerDetailBookType = {
  solo: "solo",
  series_part: "series_part",
} as const;