// apps/web/src/shared/api/generated/model/listDetailsControllerDetailTab.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type ListDetailsControllerDetailTab =
  (typeof ListDetailsControllerDetailTab)[keyof typeof ListDetailsControllerDetailTab];

export const ListDetailsControllerDetailTab = {
  all: "all",
  not_started: "not_started",
  reading: "reading",
  finished: "finished",
} as const;